@@ -2,15 +2,16 @@ package com.monke.ui;
 
 import androidx.recyclerview.widget.DiffUtil;
 
+import com.monke.rental.Rental;
+
 import java.util.List;
-import java.util.Objects;
 
-public class PhotoDiffUtilCallback extends DiffUtil.Callback {
+public class RentalDiffUtilCallback extends DiffUtil.Callback {
 
-    private final List<String> oldList;
-    private final List<String> newList;
+    private final List<Rental> oldList;
+    private final List<Rental> newList;
 
-    public PhotoDiffUtilCallback(List<String> oldList, List<String> newList) {
+    public RentalDiffUtilCallback(List<Rental> oldList, List<Rental> newList) {
         this.oldList = oldList;
         this.newList = newList;
     }
@@ -27,11 +28,11 @@ public class PhotoDiffUtilCallback extends DiffUtil.Callback {
 
     @Override
     public boolean areItemsTheSame(int oldItemPosition, int newItemPosition) {
-        return Objects.equals(oldList.get(oldItemPosition), newList.get(newItemPosition));
+        return oldList.get(oldItemPosition).getId().equals(newList.get(newItemPosition).getId());
     }
 
     @Override
     public boolean areContentsTheSame(int oldItemPosition, int newItemPosition) {
-        return Objects.equals(oldList.get(oldItemPosition), newList.get(newItemPosition));
+        return oldList.get(oldItemPosition).equals(newList.get(newItemPosition));
     }
 }
