@@ -4,6 +4,7 @@ import com.monke.rental.Rental;
 import com.monke.rental.RentalRepository;
 
 import java.util.ArrayList;
+import java.util.Calendar;
 
 import javax.inject.Inject;
 
@@ -18,10 +19,13 @@ public class PublishRentalUseCase {
     }
 
     public void execute() {
+        User user = userRepository.getCurrentUser().getValue();
         Rental rental = rentalRepository.getCreatingRental();
+        rental.setAuthorId(user.getId());
+        rental.setCreationDate(Calendar.getInstance().getTimeInMillis());
+        rental.getRealty().setAddress("");
         rentalRepository.publishRental(rental);
 
-        User user = userRepository.getCurrentUser();
         ArrayList<Rental> rentalList = new ArrayList<>(user.getRentals());
         rentalList.add(rental);
         user.setRentals(rentalList);
