@@ -1,10 +1,18 @@
 package com.monke.user;
 
+import androidx.lifecycle.LiveData;
+
 public interface UserRepository {
 
     User getUserById(String id);
 
     void saveUser(User user);
 
-    User getCurrentUser();
+    User getCreatingUser();
+
+    LiveData<User> getCurrentUser();
+
+    void setCurrentUser(User user);
+
+    void signUp();
 }
