@@ -10,12 +10,18 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.navigation.NavDeepLinkRequest;
 import androidx.navigation.fragment.NavHostFragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.monke.rental.databinding.FragmentUserRentalListBinding;
 import com.monke.rental.di.RentalComponentProvider;
+import com.monke.ui.RentalRWAdapter;
+import com.monke.ui.RentalViewHolder;
+import com.monke.user.User;
 
 import javax.inject.Inject;
 
@@ -23,6 +29,7 @@ public class RentalUserListFragment extends Fragment {
 
     private RentalUserListViewModel mViewModel;
     private FragmentUserRentalListBinding mBinding;
+    private RentalRWAdapter adapter;
 
     @Inject
     public RentalUserListViewModel.Factory factory;
@@ -45,11 +52,39 @@ public class RentalUserListFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        mBinding.btnNew.setOnClickListener(v -> {
-            NavDeepLinkRequest request = NavDeepLinkRequest.Builder
-                    .fromUri(Uri.parse(getString(com.monke.ui.R.string.new_rental_deeplink)))
-                    .build();
-            NavHostFragment.findNavController(this).navigate(request);
+        initRentalRW();
+        initNewRentalBtn();
+        observeUser();
+    }
+
+    private void initRentalRW() {
+        adapter = new RentalRWAdapter();
+        adapter.setOnItemClickListener(rental -> {
+
+        });
+
+        RecyclerView recyclerView = mBinding.listRental;
+        recyclerView.setLayoutManager(new LinearLayoutManager(
+                getContext(),
+                LinearLayoutManager.VERTICAL,
+                false
+        ));
+        recyclerView.setAdapter(adapter);
+    }
+
+    private void observeUser() {
+        mViewModel.user.observe(getViewLifecycleOwner(), user -> {
+            adapter.setRentalList(user.getRentals());
         });
     }
+
+    private void initNewRentalBtn() {
+        mBinding.btnNew.setOnClickListener(v -> {
+            NavHostFragment
+                    .findNavController(this)
+                    .navigate(R.id.action_rentalUserListFragment_to_nav_new_rental);
+        });
+    }
+
+
 }
